+++ Application/FileConverter/Windows/MainWindow.xaml.cs
@@ -13,6 +13,7 @@ namespace FileConverter
     {
         private DiagnosticsWindow diagnosticsWindow;
         private SettingsWindow settingsWindow;
+        private UpgradeWindow upgradeWindow;
 
         public MainWindow()
         {
@@ -21,6 +22,12 @@ namespace FileConverter
             Application application = Application.Current as Application;
 
             this.ConverterJobsList.ItemsSource = application.ConvertionJobs;
+
+            if (application.HideMainWindow)
+            {
+                this.Hide();
+            }
+
             if (application.Verbose)
             {
                 this.ShowDiagnosticsWindow();
@@ -28,15 +35,19 @@ namespace FileConverter
 
             if (application.ShowSettings)
             {
-                this.Hide();
-
                 this.ShowSettingsWindow();
                 this.settingsWindow.OnSettingsWindowHide += this.SettingsWindow_Closed;
             }
         }
 
         public event PropertyChangedEventHandler PropertyChanged;
-        
+
+        public void OnNewVersionReleased(UpgradeVersionDescription upgradeVersionDescription)
+        {
+            this.ShowUpgradeWindow();
+            this.upgradeWindow.VersionDescription = upgradeVersionDescription;
+        }
+
         [NotifyPropertyChangedInvocator]
         protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
         {
@@ -93,6 +104,22 @@ namespace FileConverter
             Application application = Application.Current as Application;
             application?.CancelAutoExit();
             this.settingsWindow.Show();
+        }
+
+        private void ShowUpgradeWindow()
+        {
+            if (this.upgradeWindow != null && this.upgradeWindow.IsVisible)
+            {
+                return;
+            }
+            else
+            {
+                this.upgradeWindow = new UpgradeWindow();
+            }
+
+            Application application = Application.Current as Application;
+            application?.CancelAutoExit();
+            this.upgradeWindow.Show();
         }
     }
 }